rust: Move API resources into separate files
… to prepare for these being replaced by automatically generated files.
